7.4 include mixins/videoBlock and Video Detail Edit Delete

+++ controller/videoController.js
@@ -4,8 +4,7 @@ import Video from "../models/Video";
 
 export const home = async (req, res) => {
     try {
-        const videos = await Video.find({});
-        //console.log(videos);  
+        const videos = await Video.find({}); 
         res.render('home', { pageTitle: 'Home', videos });
     } catch(error) {
         console.error(error);
@@ -39,10 +38,13 @@ export const postUpload = async (req, res) => {
     const newVideo = await Video.create({
         fileUrl: path,
         title,
-        description
+        description,
+        creator: req.user.id
     });
-    console.log(newVideo)
+    console.log(newVideo);
     // upload and save video
+    req.user.videos.push(newVideo.id);
+    req.user.save();
     res.redirect(routes.videoDetail(newVideo.id));
 };
 
@@ -52,8 +54,8 @@ export const videoDetail = async (req, res) => {
         params: {id}
     } = req;
     try {
-        const video = await Video.findById(id);
-        //console.log(video);
+        const video = await Video.findById(id).populate("creator");
+        console.log(video);
         res.render("videoDetail", { pageTitle: video.title, video });
     } catch (error) {
         console.error(error);
@@ -67,7 +69,11 @@ export const getEditVideo = async (req, res) => {
     } = req;
     try {
         const video = await Video.findById(id);
-        res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+        if(video.creator.toString() !== req.user.id){
+            throw Error();
+        } else {
+            res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+        }
     } catch (error) {
         res.redirect(routes.home);
     }
@@ -92,9 +98,14 @@ export const deleteVideo =  async (req, res) => {
         params : {id}
     } = req;
     try {
-        const video = await Video.findOneAndRemove({_id: id});
+        const video = await Video.findById(id);
+        if(video.creator.toString() !== req.user.id){
+            throw Error();
+        } else {
+            await Video.findOneAndRemove({_id: id});
+            fs.unlinkSync(video.fileUrl); // 폴더에 비디오 삭제하기 귀찮으니까...
+        }
         //console.log(video);
-        fs.unlinkSync(video.fileUrl); // 폴더에 비디오 삭제하기 귀찮으니까...
     } catch (error) {
         console.error(error);
     }
